Added Functionality for Adding books and Checkouts, UI changes and Some changes/added functionality for the backend Controllers/Services/repo

+++ src/main/java/com/cgi/library/service/BookService.java
@@ -2,6 +2,7 @@ package com.cgi.library.service;
 
 import com.cgi.library.entity.Book;
 import com.cgi.library.model.BookDTO;
+import com.cgi.library.model.BookStatus;
 import com.cgi.library.repository.BookRepository;
 import com.cgi.library.util.ModelMapperFactory;
 import org.modelmapper.ModelMapper;
@@ -25,7 +26,7 @@ public class BookService {
     }
 
     public BookDTO getBook(UUID bookId) {
-        Book book = bookRepository.getOne(bookId);
+        Book book = bookRepository.getOne((bookId));
         return modelMapper.map(book, BookDTO.class);
     }
 
@@ -34,9 +35,22 @@ public class BookService {
     }
 
     public void deleteBook(UUID bookId) {
+        if(bookRepository.findBookById((bookId)).getStatus() != BookStatus.AVAILABLE) {
+            throw new RuntimeException("Unable to Delete! Book is still checked out!");
+        }
         bookRepository.deleteById(bookId);
     }
+
     public BookDTO getBookByTitle(String title) {
         return modelMapper.map(bookRepository.findBookByTitle(title.trim()), BookDTO.class);
+    }
+
+    public Page<BookDTO> getBooksByTitleContaining(String title,Pageable pageable) {
+        return bookRepository.findBooksByTitleContainingIgnoreCase(title.trim(), pageable).map(book -> modelMapper.map(book, BookDTO.class));
+    }
+
+
+    public Page<BookDTO> getBooksByStatus(BookStatus status,Pageable pageable) {
+        return bookRepository.findBooksByStatus(status,pageable).map(book -> modelMapper.map(book, BookDTO.class));
     }
 }
